Specify a set of random integers in the range [0, upper)

Specify a set of random integers in the range [0, upper) to sample the hyperparameters from.

The semantics of this distribution is that there is no more correlation in the loss function between nearby integer values, as compared with more distant integer values. This is an appropriate distribution for describing random seeds, for example. If the loss function is probably more correlated for nearby integer values, then you should probably use one of the "quantized" continuous distributions, such as either quniform(), qloguniform(), qnormal(), or qlognormal().

randint(upper)

Arguments

upper
An integer of the upper bound for the range of integers (exclusive).

Value

A list of the stochastic expression.

See also